Move `glm_helper.hpp`
